requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.8
future==0.18.2
gunicorn==20.0.4
idna==2.8
imageio==2.6.1
intel-openmp==2019.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numpy==1.18.1
pandocfilters==1.4.2
parso==0.6.1
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
scikit-image==0.16.2
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
torch==1.5.0
torchvision==0.6.0
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
